requirements.txt
altgraph==0.17
cx-Freeze==6.1
cycler==0.10.0
future==0.18.2
kiwisolver==1.1.0
matplotlib==3.1.2
numpy==1.18.1
opencv-contrib-python==4.1.2.30
pefile==2019.4.18
Pillow==7.0.0
py2exe==0.9.2.2
PyInstaller==3.6
pyparsing==2.4.6
python-dateutil==2.8.1
pywin32-ctypes==0.2.0
six==1.14.0
